fix duh
git-svn-id: svn+ssh://frodo/cluster/svn/projects@18654 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/include/psc_util/fmt.h
@@ -17,8 +17,10 @@
  * %PSC_END_COPYRIGHT%
  */
 
+#include <stdint.h>
+
 #define PSCFMT_HUMAN_BUFSIZ	8	/* 1000.0M */
 #define PSCFMT_RATIO_BUFSIZ	7	/* 99.99% */
 
 void	psc_fmt_human(char buf[], double);
-void	psc_fmt_ratio(char buf[], int, int);
+void	psc_fmt_ratio(char buf[], int64_t, int64_t);
